requirements.txt
streamlit-authenticator==0.2.2
Pillow==9.5.0
pickle5==0.0.11
psycopg2-binary==2.9.6
DateTime==5.2

